Revise free board test
